Update README.md
Ishmael Opoku-Acheampong is a passionate fullstack developer with a keen interest in creating innovative and user-friendly web applications. With a strong background in both front-end and back-end development, Ishmael is dedicated to writing clean, efficient code and constantly expanding his skills in emerging technologies. Outside of coding, Ishmael enjoys exploring new programming languages, contributing to open source projects, and staying up-to-date with the latest trends in the tech industry.